image upscaler v2/requirements.txt
requests
Pillow
python-dotenv
gspread
google-auth
google-auth-oauthlib